Remove legacy zedwallet; rename zedwallet-beta to wrkz-wallet
