+++ src/CMakeLists.txt
@@ -43,7 +43,6 @@ file(GLOB_RECURSE WalletApi walletapi/*)
 file(GLOB_RECURSE WalletBackend walletbackend/*)
 file(GLOB_RECURSE WalletService walletservice/*)
 file(GLOB_RECURSE WalletUpgrader walletupgrader/*)
-file(GLOB_RECURSE zedwallet zedwallet/*)
 file(GLOB_RECURSE zedwallet++ zedwallet++/*)
 
 if (MSVC)
@@ -55,7 +54,7 @@ else ()
 endif ()
 
 # Group the files together in IDEs
-source_group("" FILES $${Common} ${Config} ${Crypto} ${CryptoNoteCore} ${CryptoNoteProtocol} ${Wrkzd} ${JsonRpcServer} ${Http} ${Logging} ${Logger} ${miner} ${Mnemonics} ${Nigel} ${NodeRpcProxy} ${P2p} ${Rpc} ${Serialization} ${System} ${Transfers} ${Wallet} ${WalletApi} ${WalletBackend} ${WalletService} ${zedwallet} ${zedwallet++} ${CryptoTest} ${Errors} ${Utilities} ${WalletUpgrader} ${SubWallets})
+source_group("" FILES $${Common} ${Config} ${Crypto} ${CryptoNoteCore} ${CryptoNoteProtocol} ${Wrkzd} ${JsonRpcServer} ${Http} ${Logging} ${Logger} ${miner} ${Mnemonics} ${Nigel} ${NodeRpcProxy} ${P2p} ${Rpc} ${Serialization} ${System} ${Transfers} ${Wallet} ${WalletApi} ${WalletBackend} ${WalletService} ${zedwallet++} ${CryptoTest} ${Errors} ${Utilities} ${WalletUpgrader} ${SubWallets})
 
 # Define a group of files as a library to link against
 add_library(BlockchainExplorer STATIC ${BlockchainExplorer})
@@ -111,7 +110,6 @@ add_executable(WalletService ${WalletService} ${PG_SOURCES_OS})
 add_executable(Wrkzd ${Wrkzd} ${DAEMON_SOURCES_OS})
 add_executable(WalletApi ${WalletApi} ${WALLET_API_SOURCES_OS})
 add_executable(WalletUpgrader ${WalletUpgrader} ${WALLET_UPGRADER_SOURCES_OS})
-add_executable(zedwallet ${zedwallet} ${ZED_WALLET_SOURCES_OS})
 add_executable(zedwallet++ ${zedwallet++} ${ZED_WALLET_SOURCES_OS})
 
 if (MSVC)
@@ -121,7 +119,6 @@ if (MSVC)
     target_link_libraries(zedwallet++ ws2_32 advapi32 crypt32 gdi32 user32)
     target_link_libraries(WalletApi ws2_32 advapi32 crypt32 gdi32 user32)
     target_link_libraries(miner ws2_32 advapi32 crypt32 gdi32 user32)
-    target_link_libraries(zedwallet ws2_32 advapi32 crypt32 gdi32 user32)
     target_link_libraries(WalletUpgrader ws2_32 advapi32 crypt32 gdi32 user32)
 endif ()
 
@@ -163,7 +160,6 @@ target_link_libraries(WalletApi WalletBackend)
 target_link_libraries(WalletBackend Serialization Mnemonics Nigel cryptopp-static __filesystem Utilities SubWallets Logger Config Wallet)
 target_link_libraries(WalletService JsonRpcServer Wallet Mnemonics Errors)
 target_link_libraries(WalletUpgrader Utilities WalletBackend Common)
-target_link_libraries(zedwallet Mnemonics Wallet Utilities Errors)
 target_link_libraries(zedwallet++ WalletBackend)
 
 if (OPENSSL_FOUND)
@@ -186,12 +182,11 @@ add_dependencies(Wrkzd version)
 add_dependencies(WalletUpgrader version)
 add_dependencies(WalletApi version)
 add_dependencies(WalletService version)
-add_dependencies(zedwallet version)
+add_dependencies(zedwallet++ version)
 
 # Finally build the binaries
 set_property(TARGET Wrkzd PROPERTY OUTPUT_NAME "Wrkzd")
-set_property(TARGET zedwallet PROPERTY OUTPUT_NAME "zedwallet")
-set_property(TARGET zedwallet++ PROPERTY OUTPUT_NAME "zedwallet-beta")
+set_property(TARGET zedwallet++ PROPERTY OUTPUT_NAME "wrkz-wallet")
 set_property(TARGET WalletService PROPERTY OUTPUT_NAME "wrkz-service")
 set_property(TARGET miner PROPERTY OUTPUT_NAME "miner")
 set_property(TARGET cryptotest PROPERTY OUTPUT_NAME "cryptotest")
@@ -201,5 +196,6 @@ set_property(TARGET WalletUpgrader PROPERTY OUTPUT_NAME "wallet-upgrader")
 # Additional make targets, can be used to build a subset of the targets
 # e.g. make pool will build only Wrkzd and service
 add_custom_target(pool DEPENDS Wrkzd WalletService)
-add_custom_target(solominer DEPENDS Wrkzd zedwallet miner)
-add_custom_target(cli DEPENDS Wrkzd zedwallet)
+add_custom_target(solominer DEPENDS Wrkzd zedwallet++ miner)
+add_custom_target(cli DEPENDS Wrkzd zedwallet++)
+add_custom_target(wrkz-wallet DEPENDS zedwallet++)
