--- src/zedwallet/CommandDispatcher.h
@@ -1,12 +1,0 @@
-// Copyright (c) 2018-2019, The TurtleCoin Developers
-//
-// Please see the included LICENSE file for more information.
-
-#pragma once
-
-#include <zedwallet/Types.h>
-
-bool handleCommand(const std::string command, std::shared_ptr<WalletInfo> walletInfo, CryptoNote::INode &node);
-
-std::shared_ptr<WalletInfo>
-    handleLaunchCommand(CryptoNote::WalletGreen &wallet, std::string launchCommand, Config &config);
