--- src/zedwallet/ParseArguments.h
@@ -1,9 +1,0 @@
-// Copyright (c) 2018-2019, The TurtleCoin Developers
-//
-// Please see the included LICENSE file for more information.
-
-#pragma once
-
-#include <zedwallet/Types.h>
-
-Config parseArguments(int argc, char **argv);
